requirements.txt
flask
validators
shortuuid
python-dotenv
flask-smorest
sqlalchemy
flask-sqlalchemy
flask_apscheduler